Refactor Modules/tools.py in a more logical groups
